% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crabs}
\alias{crabs}
\title{Crabs Dataset}
\format{
A data frame with 200 rows and 8 variables:
\describe{
  \item{Type}{Species code, with `B` indicating blue and `O` indicating orange.}
  \item{sex}{Sex of the crab, with `M` indicating male and `F` indicating female.}
  \item{FL}{Size of the frontal lobe in mm.}
  \item{RW}{Rear width in mm.}
  \item{CL}{Carapace length in mm.}
  \item{CW}{Carapace width in mm.}
  \item{BD}{Body depth in mm.}
}
}
\source{
Campbell, N.A., and Mahon, R.J. (1974) A multivariate study of variation in two species of rock crab of genus Leptograpsus. Australian Journal of Zoology, 22, 417-425.
}
\description{
Measurements of 200 crabs from the genus Leptograpsus, with 50 crabs for each of two colour forms and both sexes. Measurements include the frontal lobe size, rear width, carapace length, carapace width, and body depth.
}
