% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptr.R
\name{print.pptr}
\alias{print.pptr}
\title{Prints the structure of a pptr tree.}
\usage{
\method{print}{pptr}(x, ...)
}
\arguments{
\item{x}{A \code{pptr} model.}

\item{...}{Unused.}
}
\value{
Invisibly returns the input \code{pptr} model \code{x} (unchanged).
  Called for its side effect of printing the tree structure -- the oblique
  split rules and leaf predictions -- to the console.
}
\description{
Prints the structure of a pptr tree.
}
